Include scheduled tasks in agenda
